mon: factor out SequencePaxosService from PaxosService
Factor out the parts of PaxosService that assume a sequence of versioned
states into a child class.  The parent class retains all of the basic
hooks around when paxos is active and proposing and pulling state out of
the underlying store.

This will make it possible to implement a PaxosService that is not a
sequence of states.  First up: PGMonitor!

Signed-off-by: Sage Weil <sage@inktank.com>