mon/PGMonitor: store PGMap directly in leveldb
Instead of encoding incrementals and periodically dumping the whole encoded
PGMap in leveldb, instead store everything in a range of keys, and update
between versions using transactions.

Keep a list of updated pgs and osds on each version so that we know which
keys to refresh in update_from_paxos() when moving between versions.

Caveats:
 - dirty lists need to be trimmed
 - upgrade not yet implemented

Signed-off-by: Sage Weil <sage@inktank.com>