osd/PG: add readable_until; prune it
We will use this to track what we know about the readable_until bound
for the current and prior intervals.

We can prune entries from this map if the bound has passed and either
we are active (and only care about the current interval) or if the interval
is before the first_interval_readable bound in pg_history_t.

Signed-off-by: Sage Weil <sage@redhat.com>