osd: add OSD_BITWISE_HOBJ_SORT feature
We will sort [g]hobject_t's bitwise (instead of nibblewise) if all
OSDs who have participated in peering support the feature.

Note that this means a latecomer PG who does a notify may not have
the feature while the acting set operates with the new sort order.
This should not matter since it will not participate in scrub or
backfill (where the sort order matters).

FIXME: what about last_backfill?

Signed-off-by: Sage Weil <sage@redhat.com>