os/bluestore/BlueFS: make _sync_and_flush_log smarter
If we know what event we need to wait for, only wait long enough for it
to flush.  This helps the situation where another thread flushed what we
needed, and more dirty stuff was added to log_t, but we don't need to
wait for that too for our caller to be happy.

Signed-off-by: Sage Weil <sage@redhat.com>