buffer: put data and metadata in a mempool
Note that for raw_combined we leak some metadata into the data pool.

Also, we do not account for non-raw metadata or anything else in buffer.h,
as that would pollute the ABI and public interface.

Drop the namespace ceph in buffer.cc (which serves no real purpose) as it
confuses the MEMPOOL_DEFINE_* macros (they cannot be used inside a
namespace).

Signed-off-by: Sage Weil <sage@redhat.com>