os/bluestore: ExtentMap: reshard only requested range
Only reshard extents in the requested range.

Note that the strategy for unspanning blobs changed; we now span or
unspan specific blobs at the end based on our reshard result instead
of unspanning all at the start.  This keeps the spanning id stable,
which is important because it may be referenced from another that
we aren't even looking at.

Signed-off-by: Sage Weil <sage@redhat.com>