os/bluestore: release wal throttle on io finish, before cleanup
The throttle is really about limiting deferred IO; we do not need to
actually remove the WAL record from the kv db before queueing more.  (In
fact, the txc that queues more will do the cleanup.)

This avoids a deadlock where queue_transaction blocks on the throttle, the
WAL io finishes, but the kv thread is never woken up.

Signed-off-by: Sage Weil <sage@redhat.com>