mon/PGMap: move summary information into parent PGMapDigest object
Everything summary-ish that we need to send to the mon is moved into a
parent class.  The child PGMap retains the detail.

The parent gets its own encode(), and PGMap::encode_digest() will call it
to encode just the summary info.

Squashed in here is a new num_pg_by_osd that could have been done
in a preceding patch but I did things in the wrong order.  :(

Signed-off-by: Sage Weil <sage@redhat.com>